add src files and resource folder with input file
I added all the src files that we need to get up and running for the
project. It uses the old code from the Vaccum Cleaner world. We need to
check and see if the file can be read with the new scanner, and if not,
then it should be an easy fix.